move contact to contact page
